In progress: start moving all parameter handling logic into its own module
