+++ src/databricks/sql/types.py
@@ -203,56 +203,3 @@ def _create_row(
     return row
 
 
-from enum import Enum
-class DbSqlType(Enum):
-    """The values of this enumeration are passed as literals to be used in a CAST
-    evaluation by the thrift server.
-    """
-
-    STRING = "STRING"
-    DATE = "DATE"
-    TIMESTAMP = "TIMESTAMP"
-    FLOAT = "FLOAT"
-    DECIMAL = "DECIMAL"
-    INTEGER = "INTEGER"
-    BIGINT = "BIGINT"
-    SMALLINT = "SMALLINT"
-    TINYINT = "TINYINT"
-    BOOLEAN = "BOOLEAN"
-    INTERVAL_MONTH = "INTERVAL MONTH"
-    INTERVAL_DAY = "INTERVAL DAY"
-    VOID = "VOID"
-
-class DbsqlDynamicDecimalType:
-    def __init__(self, value):
-        self.value = value
-
-
-class DbSqlParameter:
-    name: str
-    value: Any
-    type: Union[DbSqlType, DbsqlDynamicDecimalType, Enum]
-
-    def __init__(self, name="", value=None, type=None):
-        self.name = name
-        self.value = value
-        self.type = type
-
-    def __eq__(self, other):
-        return isinstance(other, self.__class__) and self.__dict__ == other.__dict__
-
-SupportedParameterType = TypeVar(
-    "SupportedParameterType",
-    DbSqlParameter,
-    str,
-    int,
-    float,
-    datetime.datetime,
-    datetime.date,
-    bool,
-    decimal.Decimal,
-    type(None),
-)
-
-ListOfParameters = List[SupportedParameterType]
-DictOfParameters = Dict[str, SupportedParameterType]
